requirements.txt
fastapi==0.70.0
uvicorn==0.16.0
keras==2.4.3
tensorflow==2.3.0
keras-retinanet==1.0.0
opencv-python==4.5.1.48
pillow==10.4.0
numpy==1.18.5
python-multipart==0.0.20
